refactor(ui): split chat renderers and styles
